Document FORCE_COLOR
Document use of FORCE_COLOR variable which forces enablement of
coloring and ignore the detection logic.

There are other similar option used like CLI_COLOR, PY_COLORS but
that one is by far the most popular option. This complements
the NO_COLOR option with missed to cover for the opposite use-case
when it was designed.