Add explanation about syntax
Add explanation about syntax.

Building a syntax for a CLI is important because
it gives a strong foundation for the CLI.

It provides the following advantages:
- Provide an intuitive guide to the CLI.
- The difference between the different versions for a CLI
can be easily extracted by executing $diff to the syntax.
- Can be utilized to check the execption on the user input.
- Etc.

This commit explains how a syntax for a CLI can be establisihed,
and the advantages.

Limitation: need more examples.

Signed-off-by: Myoung-Jun Lee <lee.myoung.joon@gmail.com>